+++ src/Command/TestCommand.php
@@ -3,10 +3,8 @@
 namespace App\Command;
 
 use App\Security\AccessChecker;
-use App\Service\RecheckSearchAccessService;
 use App\Type\AccessType;
 use Ramsey\Uuid\Rfc4122\UuidV4;
-use Ramsey\Uuid\Uuid;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
@@ -16,7 +14,6 @@ use Symfony\Component\Console\Output\OutputInterface;
 class TestCommand extends Command
 {
     public function __construct(
-        private RecheckSearchAccessService $recheckSearchAccessService,
         private AccessChecker $accessChecker
     ) {
         parent::__construct();
@@ -24,12 +21,6 @@ class TestCommand extends Command
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        //        $uuid = UuidV4::uuid4();
-        //        $this->recheckSearchAccessService->markElementToBeCheckedInFuture($uuid);
-        //        $output->writeln(sprintf(
-        //            "Marked element with uuid %s for search recheck.",
-        //            $uuid->toString()
-        //        ));
         $dataUuid = UuidV4::fromString('3a3c2f8b-d1bd-40fd-b381-82de60539c9f');
 
         $uuids = $this->accessChecker->getDirectGroupsWithAccessToElement($dataUuid, AccessType::DELETE);