+++ src/EventListener/EndCheckUserSuppliedPropertiesEventListener.php
@@ -2,13 +2,13 @@
 
 namespace App\EventListener;
 
-use App\Event\ElementPropertyWriteEvent;
+use App\Event\CheckUserSuppliedPropertiesEvent;
 use App\Exception\ClientBadRequestException;
 use App\Type\RelationElement;
 
-class EndElementPropertyWriteEventListener
+class EndCheckUserSuppliedPropertiesEventListener
 {
-    public function onElementPropertyWriteEvent(ElementPropertyWriteEvent $event): void
+    public function onCheckUserSuppliedPropertiesEvent(CheckUserSuppliedPropertiesEvent $event): void
     {
         if ($event->getElement() instanceof RelationElement && $event->hasProperty('end')) {
             throw new ClientBadRequestException(detail: 'Manually setting the end property is forbidden.');