+++ src/EventListener/StartCheckUserSuppliedPropertiesEventListener.php
@@ -2,13 +2,13 @@
 
 namespace App\EventListener;
 
-use App\Event\ElementPropertyWriteEvent;
+use App\Event\CheckUserSuppliedPropertiesEvent;
 use App\Exception\ClientBadRequestException;
 use App\Type\RelationElement;
 
-class StartElementPropertyWriteEventListener
+class StartCheckUserSuppliedPropertiesEventListener
 {
-    public function onElementPropertyWriteEvent(ElementPropertyWriteEvent $event): void
+    public function onCheckUserSuppliedPropertiesEvent(CheckUserSuppliedPropertiesEvent $event): void
     {
         if ($event->getElement() instanceof RelationElement && $event->hasProperty('start')) {
             throw new ClientBadRequestException(detail: 'Manually setting the start property is forbidden.');