+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 - Add PHP intl extension.
 - Add support for RabbitMQ.
+- Partly refactor of the event system.
+- WIP calculate users and groups with search access, store result in Elasticsearch itself.
 
 ## 0.0.7 - 2023-05-18
 ### Added